chore: update folder structure
